Add feature to rectify directory of images
This also required some re-organization of the UI. The user is now asked to choose between calibration or rectification at the very start. Calibration outputs are now being organized in a "calibration_" folder, and the user can specify directories to load and save images to/from for rectification module. Also had to link the VideoDecimator feature to the rectification module.